Integrado el desactivar y activar usuario

+++ src/app/components/grupo-ocho-cruceros/grupo-ocho-cruceros.component.ts
@@ -5,6 +5,8 @@ import { Cruiser } from '../../interfaces/cruiser';
 import { ApiService } from '../../services/api.service';
 import { environment as url } from '../../../environments/environment';
 import { LocalStorageService } from '../../services/local-storage.service';
+import Swal from 'sweetalert2';
+import { SweetAlertOptions } from 'sweetalert2';
 
 @Component({
   selector: 'app-grupo-ocho-cruceros',
@@ -124,17 +126,41 @@ export class GrupoOchoCrucerosComponent implements OnInit {
   }
 
   public getDeleteAlert(data){
-    // Si marco confirmar en la moda, quiero borrar el crucero
-    if(data['confirmed']){
-      console.log("se ejecuto");
-      this.api.deleteUrl(url.endpoint.default._delete.cruisers.delete_cruiser, [data['id']])
-        .then(response => {
-          this.deleteCruiserById(response['id']);  
-        })
-        .catch(error => {
-          
-        })
+    if(data['delete']){
+       // Si marco confirmar en la moda, quiero borrar el crucero
+      if(data['confirmed']){
+        this.api.deleteUrl(url.endpoint.default._delete.cruisers.delete_cruiser, [data['id']])
+          .then(response => {
+            this.alertStatus(200, true, data);  
+          })
+          .catch(error => {
+            
+          })
+      }
     }
+    else {
+      if(data['confirmed']){
+        let form = {
+          'id': data['id'],
+          'name': data['name'],
+          'line': data['line'],
+          'model': data['model'],
+          'loadingShipCap': data['loadingShipCap'],
+          'capacity': data['capacity'],
+          'picture': data['picture'],
+          'status': !data['status']
+        }
+        
+        this.api.putUrl(url.endpoint.default._post.cruisers.post_cruiser, form)
+          .then(response => {
+            this.alertStatus(200, false, data);
+          })  
+          .catch(error => {
+            console.log("error", error);
+          })  
+      }
+    }
+   
   }
 
   /*******************************************
@@ -185,7 +211,6 @@ export class GrupoOchoCrucerosComponent implements OnInit {
   public getLayovers(id: string, route: string){
     this.api.getUrl(url.endpoint.default._get.cruisers.get_layovers, [id])
       .then(response => {
-        console.log("RESPUESTAA", response);
         this.localStorage.setItem('cruiserRoutes', response.layovers).subscribe(data => {
           this.router.navigate(['cruisers', route.split('/')[0], route.split('/')[1]]);
         })
@@ -195,5 +220,44 @@ export class GrupoOchoCrucerosComponent implements OnInit {
           console.log("errr", er);
         }
       })
+  }
+
+  private alertStatus(statusCode: number, deleted: boolean, data: any) {
+    const config: SweetAlertOptions = {
+      // tslint:disable-next-line:max-line-length
+      title: (statusCode !== 200 ? 'Se ha producido un error' : (deleted ? 'Crucero eliminado' : 'Se cambió el estatus del crucero')),
+      type:  (statusCode === 200 ? 'success' : 'error'),
+      showConfirmButton: true
+    };
+    Swal.fire(config).then( result => {
+      if(deleted){
+        this.deleteCruiserById(data['id']);
+      }
+      else {
+        this.updateCruiserFromStatus(data);
+      }
+    });
+  }
+
+  public updateCruiserFromStatus(data: any){
+    let cruisers = this.getVariableCruisers();
+    let aux = [];
+    this.cruisers.forEach(cruiser => {
+      if(cruiser['id'] === data['id']){
+        cruiser['status'] = !cruiser['status'];
+      }
+      aux.push(cruiser);
+    })
+    this.setCruisers(aux);
+  }
+
+  public cruiserSuccessfullyCreated(){
+    let config: SweetAlertOptions = {
+      title: 'Crucero creado',
+      type: 'success',
+      showConfirmButton: false,
+      timer: 900
+    }
+    Swal.fire(config).then( result =>{});
   }
 }
